requirements.txt
aioredis==1.3.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
click==8.1.8
fastapi==0.115.7
fastapi-cache==0.1.0
h11==0.14.0
hiredis==3.1.0
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
redis==5.2.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
